+++ commerce/src/codatcommerce/models/operations/get_company_info.py
@@ -4,9 +4,7 @@ from __future__ import annotations
 import dataclasses
 import requests as requests_http
 from ..shared import companyinfo as shared_companyinfo
-from ..shared import schema as shared_schema
-from codatcommerce import utils
-from dataclasses_json import Undefined, dataclass_json
+from ..shared import errormessage as shared_errormessage
 from typing import Optional
 
 
@@ -19,21 +17,6 @@ class GetCompanyInfoRequest:
 
 
 
-@dataclass_json(undefined=Undefined.EXCLUDE)
-
-@dataclasses.dataclass
-class GetCompanyInfo409ApplicationJSON:
-    r"""The data type's dataset has not been requested or is still syncing."""
-    can_be_retried: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('canBeRetried'), 'exclude': lambda f: f is None }})
-    correlation_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('correlationId'), 'exclude': lambda f: f is None }})
-    detailed_error_code: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('detailedErrorCode'), 'exclude': lambda f: f is None }})
-    error: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('error'), 'exclude': lambda f: f is None }})
-    service: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('service'), 'exclude': lambda f: f is None }})
-    status_code: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('statusCode'), 'exclude': lambda f: f is None }})
-    
-
-
-
 
 @dataclasses.dataclass
 class GetCompanyInfoResponse:
@@ -41,10 +24,8 @@ class GetCompanyInfoResponse:
     status_code: int = dataclasses.field()
     company_info: Optional[shared_companyinfo.CompanyInfo] = dataclasses.field(default=None)
     r"""OK"""
-    get_company_info_409_application_json_object: Optional[GetCompanyInfo409ApplicationJSON] = dataclasses.field(default=None)
-    r"""The data type's dataset has not been requested or is still syncing."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    schema: Optional[shared_schema.Schema] = dataclasses.field(default=None)
+    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
+    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)