+++ commerce/RELEASES.md
@@ -607,3 +607,13 @@ Based on:
 - Speakeasy CLI 1.53.0 (2.58.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [PyPI v0.33.1] https://pypi.org/project/codat-commerce/0.33.1 - commerce
+
+## 2023-08-17 00:17:19
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.73.1 (2.84.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.34.0] commerce
+### Releases
+- [PyPI v0.34.0] https://pypi.org/project/codat-commerce/0.34.0 - commerce